- Move redux connect to a separated file. - Redirect to main after getting account.
