+++ src/components/login/index.js
@@ -1,25 +1,6 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
-import LoginForm from './login-form';
-import { accountUpdated } from '../../actions/account';
-
-/**
- * Using react-redux connect to pass state and dispatch to LoginForm
- */
-const mapStateToProps = state => ({
-  account: state.account,
-  peers: state.peers,
-});
-
-const mapDispatchToProps = dispatch => ({
-  onAccountUpdated: data => dispatch(accountUpdated(data)),
-});
-
-const LoginFormConnected = connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(LoginForm);
+import LoginForm from './loginForm';
 
 /**
  * The container component containing login
@@ -31,7 +12,7 @@ const LoginComponent = () => (
       <div className={`${grid.row}`}>
         <div className={grid['col-xs-2']}></div>
         <div className={grid['col-xs-8']}>
-          <LoginFormConnected />
+          <LoginForm />
         </div>
         <div className={grid['col-xs-2']}></div>
       </div>
